Scanner struct #refactor
Since rune scanner is a standard library interface, we can just skip the hassle with `bufio` and base our API on rune scanner. It simplifies `Scan` function, point in favor. It supports the idea od a single scanner to rule all inputs, which we should favor if we are to follow CΛ RFC, another point in favor. No drawbacks at this point and no immediate ideas of such for the future.