Another case covered
Simple solution, but might not be so simple to understand. When we hit `EOF`, we need to make sure we are in the outermost zone. If we are not, it means parens are not matching, therefore end of input came too early. Checking for presence of linked parent zone turns tests green.