Stub of supervision
Since we operate on channels, we must run supervision as a separate routine. Lets also define what events we are interested in parser's work. To define scopes, we just need three. We need to be notified when abstraction starts, when it ends, and when a variable occurs. Having this information, we should be able to build nested scope.