Identity function
Denoted as capital letter `I` is the identity function. It will be our equivalent of `nil` in Λ-calculus. Note how funny the implementation is... We don't really care about the value as long as it satisfies `Λ` interface. This allows us to safely declare `I` as empty pointer! ^_^

Bam! Returning `I` from `NewΛ` turns tests green. We need better specs now...