Suddenly, green o_O
Of course, because our tests are weak. The thing is, we can't really have any stronger tests there, since we don't have any form of comparison than through identity function. At this point, it's too early to worry with all this, let's just implemnt something simple, but not a stub, just to see where we stand.